+++ src/TimeScheduler/ITimeScheduler.cs
@@ -5,7 +5,6 @@ namespace TimeScheduler;
 /// that enables deterministic control of time during testing, and normal operation
 /// in production.
 /// </summary>
-[Obsolete("Use System.TimeProvider instead. This will allow you to upgrade seamlessly the new TimeProvider API that is part of .NET 8 upon release.")]
 public interface ITimeScheduler
 {
     /// <summary>