+++ src/TimeScheduler/Testing/TestScheduler.cs
@@ -11,7 +11,6 @@ namespace TimeScheduler.Testing;
 /// <remarks>
 /// Learn more at <see href="https://github.com/egil/TimeScheduler"/>.
 /// </remarks>
-[Obsolete("Use System.Testing.TestTimeProvider instead. This will allow you to upgrade seamlessly the new TimeProvider API that is part of .NET 8 upon release.")]
 public partial class TestScheduler : TimeProvider, ITimeScheduler, IDisposable
 {
     internal const uint MaxSupportedTimeout = 0xfffffffe;